refactor: transfer from scanner to utils
